repos/alex-oliveira/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
brapi==1.2.0
brotli==1.2.0
bs4==0.0.2
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cryptography==46.0.6
cssselect2==0.9.0
curl_cffi==0.13.0
cycler==0.12.1
distro==1.9.0
Django==6.0.3
docstring_parser==0.17.0
dotenv==0.9.9
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-aiohttp==0.1.12
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-core==1.2.22
langchain-openai==1.1.12
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
multitasking==0.0.12
narwhals==2.18.0
numpy==2.4.3
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
peewee==4.0.3
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydyf==0.12.1
pyparsing==3.3.2
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
seaborn==0.13.2
selenium==4.41.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
streamlit==1.55.0
tenacity==9.1.4
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zopfli==0.4.1
zstandard==0.25.0
